requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
numpy==1.26.4
scikit-learn==1.4.2
streamlit==1.32.0
plotly==5.20.0